косяк с диапазоном дат, выпилен дабл

+++ src/main/java/travel/Tour.java
@@ -3,17 +3,17 @@ package travel;
 import java.time.LocalDate;
 
 public class Tour {
-    private Double price;
+    private Integer price;
     private LocalDate date;
     private String url;
 
-    public Tour(Double price, LocalDate date, String url) {
+    public Tour(Integer price, LocalDate date, String url) {
         this.price = price;
         this.date = date;
         this.url = url;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
